requirements.txt
streamlit
pandas
altair
numpy
snowflake-snowpark-python